requirements.txt
ccxt
requests
python-dotenv
pandas
numpy
scikit-learn
flask
